Separated shared PFG Page base into single POM
